pyproject.toml
[project]
name = "7-tech-challenge-5"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.115.8",
    "httpx==0.28.1",
    "ipykernel>=6.29.5",
    "ipython==8.13.2",
    "matplotlib>=3.10.3",
    "mlflow>=3.1.1",
    "nltk>=3.9.1",
    "numpy>=2.2.6",
    "pandas>=2.3.0",
    "pip>=25.1.1",
    "pymongo>=4.13.2",
    "python-dotenv==1.0.1",
    "ruff>=0.12.3",
    "scikit-learn>=1.7.0",
    "sentence-transformers>=4.1.0",
    "spacy>=3.8.7",
    "streamlit>=1.46.1",
    "uvicorn==0.34.0",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "decision.ipynb"
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.9
target-version = "py310"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`) codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"
